triangulation method refactor
