+++ src/main/java/sgora/mesh/editor/interfaces/TriangulationService.java
@@ -1,0 +1,17 @@
+package sgora.mesh.editor.interfaces;
+
+import sgora.mesh.editor.model.geom.Point;
+
+public interface TriangulationService {
+
+	void createNewMesh();
+
+	void addNode(Point location);
+
+	Point findNodeByLocation(Point location);
+
+	void removeNode(Point location);
+
+	void moveNode(Point location);
+
+}
